Moves Day class into own file.
